requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
azure-functions==1.11.2
certifi==2022.6.15; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.0; python_version >= '3.6'
cryptography==37.0.4; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
oracledb==1.0.2
pycparser==2.21
pyjsend==0.2.2
requests==2.28.1
urllib3==1.26.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
